:recycle: Move LearningObjectSummary to shared
