+++ src/shared/types/index.ts
@@ -13,6 +13,10 @@ import { UserDocument } from './user-document';
 import { LearningOutcomeDocument } from './learning-outcome-document';
 import { StandardOutcomeDocument } from './standard-outcome-document';
 import { ServiceToken } from './service-token';
+import {
+  LearningObjectSummary,
+  AuthorSummary,
+} from './learning-object-summary';
 
 export {
   UserToken,
@@ -26,4 +30,6 @@ export {
   UrlDocument,
   LearningOutcomeDocument,
   StandardOutcomeDocument,
+  LearningObjectSummary,
+  AuthorSummary,
 };
